+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `From<S>` for `group_concurrent::Grouped<S, G: Default>`
 - `group_concurrent::Grouped::new`
 - `Default` for `FlattenSorted`
+- `Default` for `GroupSorted`
 
 ### Changed